requirements.txt
python-dotenv==0.18.0
requests==2.25.1
env-validate==0.3
